+++ src/ts/ui/setCodeTheme.ts
@@ -6,10 +6,10 @@ export const setCodeTheme = (codeTheme: string, cdn = Constants.CDN) => {
         codeTheme = "github";
     }
     const vditorHljsStyle = document.getElementById("vditorHljsStyle") as HTMLLinkElement;
-    const href = `${cdn}/dist/js/highlight.js/styles/${codeTheme}.css`;
+    const href = `${cdn}/dist/js/highlight.js/styles/${codeTheme}.min.css`;
     if (!vditorHljsStyle) {
         addStyle(href, "vditorHljsStyle");
-    } else if (vditorHljsStyle.href !== href) {
+    } else if (vditorHljsStyle.getAttribute('href') !== href) {
         vditorHljsStyle.remove();
         addStyle(href, "vditorHljsStyle");
     }