+++ types/index.d.ts
@@ -4,6 +4,18 @@ declare module "*.png";
 
 interface Window {
     VditorI18n: ITips;
+    hljs: {
+        listLanguages(): string[];
+        highlight(text: string, options: {
+            language?: string,
+            ignoreIllegals: boolean
+        }): {
+            value: string
+        };
+        getLanguage(text: string): {
+            name: string
+        };
+    };
 }
 
 interface IObject {
@@ -350,6 +362,10 @@ interface IUpload {
     max?: number;
     /** 剪切板中包含图片地址时，使用此 url 重新上传 */
     linkToImgUrl?: string;
+
+    /** 剪切板中包含图片地址时，使用此方法进行自定义 */
+    renderLinkDest?(vditor: IVditor, node: ILuteNode, entering: boolean): [string, number];
+
     /** CORS 上传验证，头为 X-Upload-Token */
     token?: string;
     /** 文件上传类型，同 [input accept](https://www.w3schools.com/tags/att_input_accept.asp) */
@@ -391,6 +407,9 @@ interface IUpload {
 
     /** 将上传的文件处理后再返回  */
     file?(files: File[]): File[] | Promise<File[]>;
+
+    /** 取消正在上传的文件  */
+    cancel?(files: File[]): void;
 
     /** 图片地址上传后的回调  */
     linkToImgCallback?(responseText: string): void;
@@ -435,6 +454,7 @@ interface IHljs {
     enable?: boolean;
     /** 自定义指定语言: CODE_LANGUAGES */
     langs?: string[];
+
     /** 渲染右上角菜单按钮 */
     renderMenu?(element: HTMLElement, menuElement: HTMLElement): void;
 }
@@ -501,12 +521,19 @@ interface IPreview {
     theme?: IPreviewTheme;
     /** @link https://ld246.com/article/1549638745630#options-preview-actions  */
     actions?: Array<IPreviewAction | IPreviewActionCustom>;
+    render?: IPreviewRender;
 
     /** 预览回调 */
     parse?(element: HTMLElement): void;
 
     /** 渲染之前回调 */
     transform?(html: string): string;
+}
+
+interface IPreviewRender {
+    media?: {
+        enable?: boolean;
+    };
 }
 
 type IPreviewAction = "desktop" | "tablet" | "mobile" | "mp-wechat" | "zhihu";
@@ -542,6 +569,7 @@ interface IPreviewOptions {
     renderers?: ILuteRender;
     theme?: IPreviewTheme;
     icon?: "ant" | "material" | undefined;
+    render?: IPreviewRender;
 
     transform?(html: string): string;
 
@@ -752,6 +780,9 @@ interface IOptions {
 
     /** 编辑器中选中文字后触发 */
     select?(value: string): void;
+
+    /** 编辑器中未选中文字后触发 */
+    unSelect?(): void;
 }
 
 interface IEChart {
@@ -778,6 +809,7 @@ interface IVditor {
     toolbar?: {
         elements?: { [key: string]: HTMLElement },
         element?: HTMLElement,
+        updateConfig(vditor: IVditor, options: IToolbarConfig): void,
     };
     preview?: {
         element: HTMLElement,